Starting Bus Realtime map view

+++ lib/src/models/gtfs/stop_time.dart
@@ -4,14 +4,14 @@ import 'gtfs_data.dart';
 class GTFSStopTime  {
   late final Duration arrival;
   late final int stopId;
-  late final double distanceTravel;
+  late final int distanceTravel;
   late final int tripId;
   late final int stopIndex;
 
   GTFSStopTime(Map<String, String> row) {
 
     arrival = parseDuration(row["arrival_time"]!);
-    distanceTravel = double.parse(row["shape_dist_traveled"]!);
+    distanceTravel = int.parse(row["shape_dist_traveled"]!);
     stopId = int.parse(row["stop_id"]!);
     tripId = int.parse(row["trip_id"]!);
     stopIndex = int.parse(row["stop_sequence"]!);
